Support symbolic labels for statements
Add a label directive to define symbolic labels. Symbolic labels
serve as address instruction operands.

During assembly, a label is assigned the current location value
in the rom. Labels references are either resolved during assembly
(in case of a back reference) or at the end (in case of a forward
reference).

Note : Since symbolic labels can only be used as address operands,
their resolved value is used to replace the 12 least significant bits
of the instruction.