Add a limited disassembler
Add a disassembler that turns an `Assembly` into a `Program` of
`(Assembly)Statement`s.

Currently the disassembler only supports instruction without operands
and instructions with a single address operand.

The disassembler is designed to work with a symbol table, like
the assembler, so we could eventually replace adresses with labels
during disassembly.